--- .github/workflows/firebase-hosting-merge.yml
@@ -1,21 +1,0 @@
-# This file was auto-generated by the Firebase CLI
-# https://github.com/firebase/firebase-tools
-
-name: Deploy to Firebase Hosting on merge
-'on':
-  push:
-    branches:
-      - main
-jobs:
-  build_and_deploy:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: FirebaseExtended/action-hosting-deploy@v0
-        with:
-          repoToken: '${{ secrets.GITHUB_TOKEN }}'
-          firebaseServiceAccount: '${{ secrets.FIREBASE_SERVICE_ACCOUNT_LOCAL_CIRCUIT_159306 }}'
-          channelId: live
-          projectId: local-circuit-159306
-        env:
-          FIREBASE_CLI_EXPERIMENTS: webframeworks
